+++ lib/src/apps/apps.dart
@@ -1,5 +1,6 @@
 export 'instagram.dart';
 export 'ios_app_store.dart';
 export 'mac_app_store.dart';
+export 'microsoft_store.dart';
 export 'play_store.dart';
 export 'telegram.dart';